server: Satisfy the linter
gocyclo complained (rightly) about the mutate func.
This commit splits this in to smaller mutate functions
and breaks the tests in to smaller pieces. This will make
it easier to add some negative tests in future!
